docs/cdk/constructs/hello-world-l3/requirements.txt
aws-cdk-lib==2.177.0
constructs>=10.0.0,<11.0.0
aws_solutions_constructs.aws_apigateway_lambda